requirements.txt
fastapi[standard]
gunicorn
psycopg2-binary
python-dotenv